requirements.txt
pyaudio
pynput
numpy
faster-whisper
rumps
pyobjc
py2app
boto3
python-dotenv
pyperclip